Make C_stack_pointer work with LLVM backend
When using LLVM, allocating zero bytes on the stack is legal but
is documented to return an undefined value.  In practice, this is
NULL if optimization is enabled, or alloca(1) if not.  Work around
this by using alloca(1) to get the stack pointer or, in supported
environments, using inline asm.   alloca(1) may waste a machine word,
but does not always; also, llvm uses the stack more efficiently,
so the effects tend to even out.