When referencing libchicken, take PROGRAM_PREFIX and PROGRAM_SUFFIX into account
The build system names libchicken taking PROGRAM_PREFIX and
PROGRAM_SUFFIX into account (to assemble INSTALL_LIB_NAME), but the
chicken tools don't.  `libchicken' is hardcoded in chicken tools and
tests.

This patch makes cross tools load libchicken according to the name
generated by the build system.  It also tries to keep tests working on
windows, in cases that libchicken has to be copied to the tests dir
(this is a bit hacky, since the actual lib name is not known from
scripts, so lib*chicken*.dll is copied).

This patch doesn't handle the case when TARGET_LIB_NAME is used by csc
(i.e., when not in host-mode).  TARGET_LIB_NAME is hardcoded to
`chicken'.

Signed-off-by: Peter Bex <peter.bex@xs4all.nl>