chicken-install: abort when transport and location cannot be determined
This patch fixes the number of arguments passed to the procedure given
as argument to `with-default-sources', which would raise an arity
error when transport and location could not be determined (issue found
by Peter Bex).

That procedure is now given a thunk that will properly abort
chicken-install, displaying a meaningful message.

Signed-off-by: Peter Bex <peter.bex@xs4all.nl>