Fix for star-export with explicit re-export of syntax (#882)
In a module with a "*" export list that explicitly re-exports syntax
(via "export") imported from another module, the export is currently
not externally visible. The reason for this is that the re-exported
syntax does not appear in the "sexports" slot of the module record
from the module that does the export. It is also not in the list of
syntax definition (since it is not defined in that module). To make
this work, "export" has to check for a module with "*" export-list and
add the information of the rexported syntax found in the current
macro-environment to the "sexports" slot. Finalization of the module
will merge syntax definitions found in the definition-list and in
sexports to make the reexport work with import-libraries.

Part of this patch is from sjamaan, the problem was reported by
megane.

Signed-off-by: Christian Kellermann <ckeen@pestilenz.org>