Fix computation of stack depth beyond C_stack_limit.
The old way of calculating the depth could overflow if the stack is
located very high in memory. This was causing incorrect stack-overflow
errors on some 64-bit systems (sparc64 and x86-64). The new method
checks whether the stack-pointer is beyond the limit computed at
startup, including a 64k "slack" area to allow (modestly) deep
recursions in direct-call-optimized (simple, non-allocating, not
externally visible) procedures.

Signed-off-by: Christian Kellermann <ckeen@pestilenz.org>