Do not clear eof-status of input port when reading from file-port.
Calling "(read-string #f)" or "(read-all)" from csi in a terminal will
not detect EOF, waiting forever for more input. This is caused by
"C_fast_read_string_from_file", which calls clearerr(3) to clear the
error/eof status on encountering EOF. But when the input port is
connected to a terminal, input may still follow and the EOF status
will be lost. The patch removes the call to clearerr(3) and adds
an additional EOF check at the beginning. I'm not 100% sure about
this fix, but it seems to do the job.

Signed-off-by: Christian Kellermann <ckeen@pestilenz.org>