Reduce typeset in alternative conditional branch with predicate.
During flow-analysis, when a predicate is applied to a variable, the
variable is assumed to have the corresponding type in the consequent
branch of a conditional that depends on this predicate call. This
patch adds a small enhancement that, in case the variable type is
known to be a typeset (an "(or ...)" type), reduces the typeset by
removing those types that match the predicate-type:

(let ((a ...))  ; say "a" is of type "(or string number)"
  (if (number? a)
      ...           ; "a" is known to be of type "number"
      ...))         ; "a" is now known to be of type "string"  <- new

Here "number" matches the predicate type of "number?" ("number"),
is removed from the "(or string number)" type, and results in
type "string" for "a" in the second "if" branch.

Signed-off-by: Peter Bex <peter.bex@xs4all.nl>