Fix TO argument check in subvector procedure
The subvector procedure checked its TO argument as if it was
inclusive. However, its default value is the passed vector's length so
it must be exclusive which also corresponds with the implementation. In
effect, the erroneous check prevented producing a subvector including
the given vector's last element. This patch changes the TO argument
check to be exclusive and adds some test cases for the subvector
procedure.

Signed-off-by: felix <felix@call-with-current-continuation.org>