Fix #985 by making process ports consistent with TCP ports.
This causes it to call "fetch" only when more data is requested
than available in the buffer, instead of always calling "fetch" and
checking inside the procedure whether we need more data.

The bug was due to the fact that fetch checked the position was at
the end of the buffer, but it wasn't since ##sys#scan-buffer-line
and the posix eos-handler doesn't advance the position while reading
(only afterwards, assuming "fetch" would reset the position).

Signed-off-by: Christian Kellermann <ckeen@pestilenz.org>