Avoid context switch during TCP errno reporting
There is currently the potential for a scheduler context
switch between when the global (errno) is updated and the
when the textual error message is obtained.  This can
also happen if a non-inlined procedure is called prior
to updating (errno).

We fix the first by using the return value of
(##sys#update-errno), which is the updated value,
as the message errno.  The second is fixed by avoiding
a separate call to (fail).

We also consolidate error handling into a macro, which
catches a couple instances where (errno) was not updated.

Signed-off-by: Felix Winkelmann <felix@call-with-current-continuation.org>