Use "noreturn" attribute in newer clang versions and get rid of a clang warning.
The division procedure C_2_divide would try to call C_fix on the
uninitialised variable "iresult" in the case where either of the
divident or the divisor values are flonums ("fflag" will be 1 then).

In the situation where neither value is a flonum, it'll either barf
(which is declared "noreturn"), or properly initialise iresult.

This means the C_unfix() call on an uninitialised value which gets
C_fix()ed later is pure overhead in the case either value is a flonum.

Signed-off-by: Moritz Heidkamp <moritz@twoticketsplease.de>