Git rid of endless GC loop while handling signals. Fixes #989 (and #877?)
Signal handling is intertwined with the GC for performance reasons.
When a signal is received, the real C signal handler will put the
signal on a pending stack for Scheme to pick up.  In order to get the
signal handler invoked ASAP, it will also force the stack to appear
"full" to the GC.

If a signal is received during a Scheme "signal handler", the stack
will be set to "full" in the real signal handler, and control will
then continue in the regular code, as is usual in POSIX.  The next
signal is fetched by ##sys#signal-hook through C_i_pending_interrupt,
which clears interrupt_reason.  Then it calls the inner LOOP in
##sys#signal-hook, which performs a stack_probe in the generated C
code which will cause a minor GC.

Thus, the stack is "full" but handle_interrupt() is not reached (as
interrupt_reason was already cleared).  So the GC will clear the stack
but it cannot because the stack is mostly empty already.  This means
it will still seem "full" after the GC has finished.  Then, the
trampoline calls proc, which again does a stack_probe which triggers a
minor GC, ad infinitum.

To make things worse, this only happens if the user-mode signal
handler is uninterruptible (ie, native code or code with a
disable-interrupts declaration).  If the user code contains its own
stack_probe it will get to it before ##sys#interrupt-hook has a chance
to clear interrupt_reason, thus it will cause the regular handler to
be invoked through save_and_reclaim().  In order to normalise all this
behaviour, the route into handle_interrupts() through the GC is now
blocked while ##sys#interrupt-hook is running.  This block also
applies to forcing the stack to appear full while handling
interrupt-hook as that's not required because we're still looping and
fetching interrupts from the queue.

Signed-off-by: Mario Domenech Goulart <mario.goulart@gmail.com>