Remove HAVE_GCVT check and definition. Use snprintf on all platforms.
gcvt() is deprecated in POSIX, and it produces problematic output in
at least cygwin.  For example, -0.0 is printed as "0.0" and more fun
things.  This fix adds a regression test for the negative zero case.

Replace sprintf usage with snprintf just in case someone decides to
use insane values for flonum_print_precision.

Signed-off-by: Moritz Heidkamp <moritz@twoticketsplease.de>