Go back to sleep when thread-join! is called without timeout.
This patch fixes an issue discovered by Michael Greenly. When a signal
handler is called a thread waiting for another with thread-join! got
woken up and the code assumed this could have happened only because
the other thread died or the timeout occured. Hence if the waited-for
thread is not in state terminated or dead a timeout exception is
thrown.

With this patch the thread is put back to blocking state (for
termination of the waited-for thread) if no timeout has been given.

Note: This is reliably triggered only when the signal is delivered
external from the CHICKEN process.

The patch also refactors the code to explicitly match the expected
thread states and errors out in the else clause.

A test case for this situation has been added.

Signed-off-by: Peter Bex <peter.bex@xs4all.nl>