Fix race condition in #877.
Don't allow a flood of identical signals to hog the signal queue.

In the test, just to be sure, first set up a signal that the child can
send the parent when it's ready.  Then start sending signals to the child,
so we know the child managed to set up all its handlers.

Finally, wait for acknowledgement of child shutdown so we can detect
whether the error persists.

Signed-off-by: Christian Kellermann <ckeen@pestilenz.org>