Prevent problems when the host system does provide strlcpy/strlcat.
Defining strlcat/strlcopy causes trouble if HAVE_STRLCAT/HAVE_STRLCPY
is not set and the host system does offer these functions.  This could
also happen on systems that don't offer these functions but
bsd/string.h is being included by user code on Linux, for example.

Instead of defining strlcat and defining C_strlcat to point to that,
we define the replacement function directly under the name C_strlcat.
On systems which do have strlcat, we define C_strlcat as an alias of
strlcat.

Signed-off-by: Mario Domenech Goulart <mario.goulart@gmail.com>