Fix bug #1106 chicken-install -deploy dependencies
This commit will make chicken-install ignore the local egg db while
traversing the dependency list when in deployment mode. This results
in all dependencies installed in the deployment directory, instead of
just the explicitly stated one in the egg's meta file.

I have tested this with a spiffy example and

chicken-install -deploy -p $(pwd)/t spiffy

Which installs all dependencies of spiffy and intarweb.

Signed-off-by: Peter Bex <peter.bex@xs4all.nl>