Fix #878 which was indeed a bug, caused by an incorrect hand-rolled optimisation.
This adds some more "integration" test cases so that we can verify
more easily that the combination of make-kmp-restart-vector and
kmp-step is looping through the pattern correctly as it advances
through the search string.

The optimisation was due to a mistaken reading of the reference
implementation: the pattern was indexed as pat[k+1] at the j=-1 case
and as pat[k] at the pat[k]=pat[j+start] case, but the optimisation
changed the code to use pat[k] in both cases.

Signed-off-by: Mario Domenech Goulart <mario.goulart@gmail.com>