Get rid of explicit marking of primitives re-exported by modules.
When importing a module, its dependencies always *must* be imported,
because a re-export could involve a macro which is defined by another
module, and the definitions of macros are handled through the import
library.

Because this is happening anyway, any module re-exporting a primitive
will end up importing the core module which exports it.  This import
will execute ##sys#register-primitive-module which causes the primitives
to be marked though ##sys#primitive-alias.

Because of this recursive module loading, it's unnecessary to detect
primitives at module registration time.  This simplifies the module
registration code a little (which is already hairy enough of itself).

Signed-off-by: Evan Hanson <evhan@foldling.org>