Remove ##sys# prefix from lambda-info names of library procedures
This is a cosmetic change that removes the "##sys#" prefix from the
lambda-info names of procedures defined in library.scm. Where a
procedure was defined first with the prefix and later as an alias
without it, their definitions have been swapped, making sure the
non-prefixed name is used in its lambda-info structure.

Signed-off-by: Christian Kellermann <ckeen@pestilenz.org>