Move foldable binding annotations into types.db
This adds a #:foldable property for procedures in types.db and migrates
the list of foldable bindings out of c-platform.scm and into that file.

It also makes the optimizer consider #:predicate procedures foldable,
unmarks some identifiers that shouldn't be marked foldable, and adds a
handful of identifiers from the core units that should.

Also, update the list of standard and extended bindings in manual/faq
and remove all remaining references to hash-table-ref, thread-specific,
and thread-specific-set!

Signed-off-by: Peter Bex <peter.bex@xs4all.nl>