Avoid invalid specializations for multi-valued foreign-primitives
The foreign-primitive form may return multiple values with the C_values
function, but its expansion is always declared to return a (single)
undefined value. This triggers invalid ##sys#c-w-v specializations for
single-valued producers where multiple values may in fact result.

This fixes that declaration so that multiple values are expected when no
return type is specified.

Signed-off-by: Peter Bex <peter.bex@xs4all.nl>