Fix incorrect optimization in cond expansion.
When a constant is used as the condition, the consequence would be
simply taken as-is (or wrapped in a "begin" if it's multiple
statements).  However, if there's a => following the constant, it should
expand to a call of the procedure with the constant as argument.

Reported by Michele La Monaca

Signed-off-by: Evan Hanson <evhan@foldling.org>