Fix buffer overrun in substring-index[-ci]
When passing a start index greater than 0, substring-index[-ci] would
scan past the end of the subject string, leading to bogus results in
case the substring is accidentally run into beyond the end of the
subject. This patch fixes the issue and also adds a range check for the
start index.

Signed-off-by: Peter Bex <peter.bex@xs4all.nl>