Introduce a new transient memory region: the "scratch space", and update integer abs and integer negate to use it.
This memory region is used when a variable amount of memory is needed
in an inline function.  We don't have a continuation to save in these
situations, so we set aside the variable part of the object in the
scratch space, but the returned object itself must still be allocated
on the stack.  This can only be done if the returned object is
"immutable".  This is required because we need to update the object
slot when the scratch space is resized and the data part is moved.

The stack-allocated object's slot pointer is remembered in a pointer
which directly precedes the stored object in memory.  This allows us
to update the slot without having to do a liveness analysis while
resizing the scratch space; we can just update the one slot and move
the object with impunity.

The scratch space is reclaimed at the earliest opportunity, which will
be during a minor GC.  This ensures we don't get runaway memory
allocation situations.  Because some part of the object is still
stored on the stack, even in a tight loop that gets compiled to a C
loop, there will still be stack built up, ensuring an eventual GC.

The GC is extended to understand that the scratch region may(!) exist,
and will copy objects from it into the heap, like stack-allocated
objects, making it a proper extension of the nursery.

Currently, the only objects stored like this are (of course!) bignums,
but this approach could also be used to allocate u8vectors, for
example: they have a similar structure, with a wrapper struct pointing
to bytevector/blob data.