Convert gcd to use scratch space, making it inlineable.
This is in preparation of converting + and - to use scratch space,
since those need gcd support when dealing with ratnums.  Unfortunately,
by making gcd inlineable, we can't (yet) make use of Burnikel-Ziegler
division, which means that gcd on large bignums will not benefit from
the fancy algorithms.

We drop Lehmer's GCD implementation (at least for now), because
transcribing it to C using the current API would result in many
more piles of hairy code.

Because the inlineable GCD does not receive a continuation, yet
still needs to create an unbounded number of intermediate bignums,
this requires a way to "re-use" a fixed number of buffers.  We add
a new "migrate" function which can move all components of an object
that live in a particular buffer into another buffer.  If the
object does not live in that buffer (or is immediate), it will not
be touched.  This allows us to drag along the x and y objects in
gcd with only one "lookbehind buffer" that *may* hold the previous
iteration's intermediate object.  On the first iteration, x and y
will be provided by the caller and hence not live in the buffer, so
these won't be touched.