Convert dyadic integer multiplication (including Karatsuba!) to use scratch space.
The multiplication code tries to run Karatsuba only if there's enough
room on the stack.  This means the runtime performance may be somewhat
unpredictable, but at least it allows us to still have a more-or-less
readable version of Karatsuba, which is a fundamentally recursive
algorithm.  If anyone feels up to it, this could be improved by making
it iterative with a dynamically allocated stack of TODOs, making the
algorithm use constant C stack space, ensuring it can always be used.

And, best of all, it means multiplication can be inlineable.