Simplify division of bignums and remove a lot of boilerplate code.
Unfortunately, this gets rid of Burnikel-Ziegler division (for the
time being)

The Burnikel-Ziegler optimization wasn't as effective as Karatsuba in
the first place, and it got in the way of inlineability for quotient
and remainder.  And besides, if it isn't too difficult we can restore
it later in a similar way like Karatsuba multiplication.