Fix potential buffer overrun error in string-translate*
string-translate* would scan from every position in the target string
for each source string in the map, even if that would mean scanning
past the end.  The out-of-bounds read would be limited to the size of
the overlapping prefix in the trailing garbage beyond the string,
because memcmp will stop scanning as soon as there is a different
byte in either of the memory areas.

This also adds a few basic tests for string-translate*

Signed-off-by: Evan Hanson <evhan@foldling.org>