Fix create-directory parent dir creation on Windows.
When passing #t as the second argument to make create-directory behave
like "mkdir -p", on Windows there was a small mistake in the logic so it
would never actually create the topmost parent directory, only those at
level 2 and below.  This was exposed by the find-files test which uses
this feature of create-directory.

Instead of having differing implementations, we move the UNIX
implementation into posix-common; it recursively decomposes pathnames
using standard procedures that already deal with the difference in path
separator.  Both use C_mkdir(), which is defined in a platform-specific
way (but using a common API) at the top of each corresponding platform's
posix file.

Signed-off-by: Evan Hanson <evhan@foldling.org>