Port B/Z division perf stability fix from numbers.
This changes Burnikel-Ziegler division's threshold check to look at
the *difference* between the two arguments instead of the size of the
smallest one.  This idea is taken from MCA, algorithm 1.9.

This seems to make a big difference on the "pidigits" benchmark from
the Great Language Shootout game, which somehow triggers such
pathological behaviour of the B/Z division routine that using plain
gradebook division would be up to three times faster on that
benchmark.  Other benchmarks are mostly unaffected by this issue.

Many thanks to "balkenbrij" on Reddit for finding this issue!