Fix normalize-pathname for dot-relative paths beginning with ".//"
Because dots aren't pushed onto the accumulated list of pathname parts,
`normalize-pathname` would hit the second slash, see that there were no
leading parts, and consider the path absolute.

To fix this, we make the function save the type of the path, either
'abs(olute) or 'rel(ative), as soon as it's known, and avoid overwriting
it from then on.

Fixes #1202.

Signed-off-by: Peter Bex <peter@more-magic.net>