Fix unsafe specializations in types.db
This patch fixes some specializations in types.db which could lead to
unsafe code. In all cases, the specialized versions did not only elide
runtime type checks but also range checks for their arguments. For
example, `string-ref' could have been specialized so that it would allow
for an index pointing past the end of the string to be passed.

Fixes #1216.

Signed-off-by: Peter Bex <peter@more-magic.net>