Standardize specialization and argument type matching
Removes the concept of "exact" matching, making the behaviour of
specializations and `compiler-typecase` more like that of normal
flow-analysis. This makes it possible to specialize on implicit union
types such as `number` where previously such specializations would never
be triggered (because, for example, `number` would never match `fixnum`
or `flonum` "exactly").

Ensures that user-defined specializations take precedence over built-in
ones, and that specializations are prioritized by the order in which
they're defined.

Refactors `match-types` slightly in order to remove some redundant code
and standardize idioms, and adds a handful of scrutinizer tests.

Fixes #1214.

Signed-off-by: Peter Bex <peter@more-magic.net>