Try to detect corrupted data before performing GC_REALLOC.
If an "impossible" object in the stack or heap: one that has a size
that's larger than the memory area that contains it, we know we have a
data corruption on our hands.  Panic immediately instead of going in
for the reallocating GC.  This prevents the memory state from being
mutated any more, which may help in debugging.

Signed-off-by: Evan Hanson <evhan@foldling.org>