Mark external type declarations as declared.
By not being marked as "declared", types loaded from a types database
would be considered to be inferred via flow analysis.  When scrutinizing
procedure definitions, "initial-argument-types" and "variable-result"
would simply return '* or '(*) as the type, which doesn't match the
loaded declaration.  This had the effect of blocking specialization.

Because CHICKEN 4 doesn't use modules for its core procedures, we will
only do this when building CHICKEN itself.  User code may define toplevel
procedures which match names from core, if the matching units are not
loaded this is okay, but we can't apply the specializations in that case.

This fixes the most important part of #1219.

Signed-off-by: Evan Hanson <evhan@foldling.org>