Add simple statistical profiler to runtime library
This enables collection of profiling data via statistical sampling to
every program built with CHICKEN.  It relies on trace information for
determining which procedure is running.  This also means it has a finer
granularity than the default instrumentation-based profiler.  This can
be an advantage or disadvantage depending on what you're trying to do.

Signed-off-by: Evan Hanson <evhan@foldling.org>