Fix UWORD_COUNT_FORMAT_STRING on 64bit platforms.
It should be "unsigned long", not just "unsigned integer".

This allows us to remove some questionable casts.
heap{space[12]}_size are also changed from size_t to C_uword because
all the intermediate calculations for them are done through C_uwords.
Besides, C_maximal_heap_size is a C_uword so it doesn't need to be
size_t because the value of heap_size can never be bigger than
C_maximal_heap_size.  This also helps with printf for heap_size,
because the "z" modifier seems not as widely supported as %lu.

Signed-off-by: Evan Hanson <evhan@foldling.org>