Use pair as loop result handle in compiler-syntax for `map`
This allows the expansion to blindly dereference the second slot of the
value accumulated by the loop, avoiding the need for a conditional that
detects the first iteration and updates the loop's result box and
removing the continuation resulting from that conditional. Because the
map-loop is then free of any CPS calls, it compiles to a label and goto.

Signed-off-by: Peter Bex <peter@more-magic.net>