Fix unsafe code generation for procedure calls.
When the debugger was integrated, ##core#call in the *target* language
was slightly changed: it got an additional leading argument that
indicates the debug info index, so the "safe-flag" option was moved to
position number 2.

However, the code generation would still interpret the first position
as "safe-flag", which meant C_fast_retrieve_proc() was omitted even in
cases where the called object was not guaranteed to be a closure.
C_fast_retrieve_proc will check the object and raise an exception if
it's not a closure.

Signed-off-by: felix <felix@call-with-current-continuation.org>