Preserve type of argument in car in assq/assv procedure results.
If the result is a pair, the car will always be the same type as the
first procedure argument.  The list may consist of pairs of union
types or unknown types, so this allows us to recover some type info
due to the first argument in most cases being more specific.

Signed-off-by: Evan Hanson <evhan@foldling.org>