Fix module db generation on Windows.
On Windows, the generated pathname for the glob in update-db would
consist of slashes, but the final directory name would contain a
backslash between the binary version and the filename.  The regular
expression which extracted the module name from the full path to the
import file only matched on slashes, causing the backslash to be
included in the module name.

Instead of regexes, we now use platform-aware path manipulation
procedures which honour the system's separator character(s).

Signed-off-by: Evan Hanson <evhan@foldling.org>