Simplify code generation by only using C_demand()
This obsoletes C_stack_probe(), which was a bit broken in general,
because it was passed the base pointer.  This would only work reliably
on architectures where the stack grows downward; on architectures where
the stack grows upward, we would have to add the allocated amount to the
base pointer to find out how much stack was used.

In many cases this didn't break because often we statically allocate
using the "C_word *ab[DEMAND]; C_word *a = ab;" idiom.  In this case,
"a" would be allocated at the end of the stack, before/after the entire
buffer.  But in situations where we dynamically allocated using the
"C_word *a; a = C_alloc(DEMAND);" idiom, "a" might still fall within
range of the stack, while a+DEMAND might push it just over the stack
reserve.

Besides simplifying and removing a potential bug on platforms where the
stack grows upwards, this change makes it possible to add argvector
allocation to the C_demand() call.

Signed-off-by: Evan Hanson <evhan@foldling.org>