Fix invalid base handling of string->number
Before, we'd segfault when given absurd base arguments like 60,
depending on how the libc implemented this undefined behaviour.  Now we
detect it early and bail out.

The handling of invalid values is slightly improved as well; we might
get a return value of 0 from strtod and strtol(l) on error: HUGE_VAL /
(LONG_)LONG_{MAX,MIN} are only returned on overflow/underflow, and the
errno might have a different value from ERANGE as pointed out by the
glibc manual page, even though the spec is silent about other errnos.
At least if the string couldn't be converted eptr will be mutated to
point to str.  But otherwise, all bets are off.

The spec for strtod() even says:

 "If the result underflows (7.12.1), the functions return a value whose
 magnitude is no greater than the smallest normalized positive number in
 the return type; whether errno acquires the value ERANGE is
 implementation-defined"

Attempting to write C code: not even once...

Fixes #1272.

Signed-off-by: Evan Hanson <evhan@foldling.org>