Fix #1274 by grouping mvars and vars together.
Instead of collecting mvars and vars in separate variables, which means
we lose their correct ordering, we now store them in "vars" and their
values in "vals".  We still keep a separate "mvars" list around which
holds #t and #f to distinguish mvars from non-mvars, because non-mvars
are implicitly MV; additional values after the first must be silently
ignored.  Also, set! is more efficient than call-with-values for one
argument.

Signed-off-by: Evan Hanson <evhan@foldling.org>