Avoid triggering stack overflows in signal handler
The signal handler would touch C_stack_limit in order to force a GC, but
this only really works for C_demand().  The C_stack_limit is also
checked by C_stack_check1, but this check is used to detect runaway C
recursion and *actual* stack overflow errors.

This patch introduces a distinction between the C_stack_limit of
old (which is unchanged in meaning) and a new C_stack_hard_limit, which
is the true stack limit.  The C_demand checks are still against
C_stack_limit, but C_stack_check1 checks are against C_stack_hard_limit.

Fixes #1283

Signed-off-by: Evan Hanson <evhan@foldling.org>